requirements.txt
djangorestframework
django-filter
// No additional dependencies needed.